+++ src/containers/Settings.js
@@ -3,27 +3,21 @@
  * @flow strict-local
  */
 import React from 'react';
-import {StyleSheet, View} from 'react-native';
+import {View} from 'react-native';
 import {connect} from 'react-redux';
-import Profile from '../containers/Profile';
-import Languages from './Languages';
+import LanguageSelection from './LanguageSelection';
+import SettingsRow from '../components/SettingsRow';
 
 import {mapDispatchToProps} from '../ducks/actions';
 
 const Settings: () => React$Node = props => {
   return (
-    <View style={styles.view}>
-      <Profile />
-      <Languages />
+    <View>
+      <SettingsRow title={'User ID'} content={props.userID} />
+      <LanguageSelection />
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  view: {
-    flex: 1,
-  },
-});
 
 const mapStateToProps = state => state;
